refact concurrency
previously, each iteration create a thread to process a request, but
most of those threads were only waiting for a request and only one of
those threads processed it.

now in the loop we wait for the request to arrive and then process it in
the thread-pool.

this way we only add a task to the thread pool when we get a request.

we also use async/await to optimize threads.